backend/requirements.txt
fastapi[standard]>=0.110.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
